+++ frontend/src/metabase/search/components/SearchResult/SearchResult.styled.tsx
