+++ frontend/src/metabase/search/components/SearchResult/SearchResult.unit.spec.tsx
