+++ frontend/src/metabase/search/components/SearchResult/components/Score.tsx
@@ -1,0 +1,7 @@
+import type { SearchScore } from "metabase-types/api";
+
+export function Score({ scores }: { scores: SearchScore[] }) {
+  return (
+    <pre className="hide search-score">{JSON.stringify(scores, null, 2)}</pre>
+  );
+}
