+++ frontend/src/metabase/search/components/SearchResult/components/TableIcon.tsx
@@ -1,0 +1,5 @@
+import { Icon } from "metabase/core/components/Icon";
+
+export function TableIcon() {
+  return <Icon name="database" />;
+}
