+++ frontend/src/metabase/search/components/SearchResult/components/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_ICON_SIZE = 16;
