+++ frontend/src/metabase/search/components/SearchResult/components/index.ts
@@ -1,0 +1,7 @@
+export { CollectionIcon } from "metabase/search/components/SearchResult/components/CollectionIcon";
+export { Context } from "metabase/search/components/SearchResult/components/Context";
+export { DefaultIcon } from "metabase/search/components/SearchResult/components/DefaultIcon";
+export { ItemIcon } from "metabase/search/components/SearchResult/components/ItemIcon";
+export { Score } from "metabase/search/components/SearchResult/components/Score";
+export { TableIcon } from "metabase/search/components/SearchResult/components/TableIcon";
+export * from "./constants";
