+++ frontend/src/metabase/search/components/SearchResult/index.ts
@@ -1,0 +1,1 @@
+export { SearchResult } from "./SearchResult";
