+++ frontend/src/metabase/search/containers/SearchResultSection.tsx
@@ -1,0 +1,11 @@
+import type { WrappedResult } from "metabase/search/types";
+import Card from "metabase/components/Card";
+import { SearchResult } from "metabase/search/components/SearchResult/SearchResult";
+
+export const SearchResultSection = ({ items }: { items: WrappedResult[] }) => (
+  <Card className="pt2">
+    {items.map(item => {
+      return <SearchResult key={`${item.id}__${item.model}`} result={item} />;
+    })}
+  </Card>
+);
